feat: implement auth check on SplashScreen and add activity indicator
This commit introduces an authentication check on the SplashScreen. After the splash animations complete, the app now checks AsyncStorage for a `userToken`.

If a token is found, the user is navigated to 'NavigationBottom'. Otherwise, they are navigated to 'Login'.

An `ActivityIndicator` has also been added to the SplashScreen to provide visual feedback during this process. `useNavigation` hook is now used for navigation.